fix some swig warnings

+++ numerics/src/LCP/lcp_pivot.c
@@ -267,10 +267,10 @@ void lcp_pivot_covering_vector(LinearComplementarityProblem* problem, double* re
   {
     /* Principal Pivoting Methods  */
     case SICONOS_LCP_PIVOT_BARD:
-      basis[block] = basis[block] <= dim ? block + dim + 2 : block + 1;
+      basis[block] = basis[block] <= (int)dim ? block + dim + 2 : block + 1;
       break;
     case SICONOS_LCP_PIVOT_LEAST_INDEX:
-      basis[block] = basis[block] <= dim ? block + dim + 2 : block + 1;
+      basis[block] = basis[block] <= (int)dim ? block + dim + 2 : block + 1;
       break;
     case SICONOS_LCP_PIVOT_PATHSEARCH:
       DEBUG_PRINTF("t value : %le\n", mat[t_indx]);
@@ -381,7 +381,7 @@ void lcp_pivot_covering_vector(LinearComplementarityProblem* problem, double* re
     }
 
     DEBUG_PRINTF("driving variable %i \n", drive);
-    if (basis[block] == dim + 1)
+    if (basis[block] == (int)dim + 1)
     {
       if (pivot_selection_rule != SICONOS_LCP_PIVOT_PATHSEARCH)
       {
@@ -436,10 +436,10 @@ void lcp_pivot_covering_vector(LinearComplementarityProblem* problem, double* re
     {
       /* Principal Pivoting Methods  */
       case SICONOS_LCP_PIVOT_BARD:
-        basis[block] = basis[block] <= dim ? block + dim + 2 : block + 1;
+        basis[block] = basis[block] <= (int)dim ? block + (int)dim + 2 : block + 1;
         break;
       case SICONOS_LCP_PIVOT_LEAST_INDEX:
-        basis[block] = basis[block] <= dim ? block + dim + 2 : block + 1;
+        basis[block] = basis[block] <= (int)dim ? block + (int)dim + 2 : block + 1;
         break;
       case SICONOS_LCP_PIVOT_PATHSEARCH:
         leaving = basis[block];
